Implement time-range delta determination tests.
